add in ability to fire torpedos from the ship

+++ src/Sprite.h
@@ -15,20 +15,20 @@
 
 class Sprite : public Renderable {
     TextureRef m_texture;
-    Vector2i m_position;
-    Vector2i m_size;
-    Vector2i m_center;
+    Vector2f m_position;
+    Vector2f m_size;
+    Vector2f m_center;
 public:
     Sprite(TextureRef m_texture);
     virtual void render(Renderer& rend);
 
-    void set_position(const Vector2i& position) { m_position = position; }
-    const Vector2i& position() const { return m_position; }
+    void set_position(const Vector2f& position) { m_position = position; }
+    const Vector2f& position() const { return m_position; }
 
-    void move_position_by(const Vector2i& adjust_position);
+    void move_position_by(const Vector2f& adjust_position);
 
-    void set_center(const Vector2i& center) { m_center = center; }
-    const Vector2i& center() const { return m_center; }
+    void set_center(const Vector2f& center) { m_center = center; }
+    const Vector2f& center() const { return m_center; }
 
-    const Vector2i& size() const { return m_size; }
+    const Vector2f& size() const { return m_size; }
 };
